chore: move files into folders
