+++ src/App.test.tsx
@@ -6,6 +6,6 @@ import '@testing-library/jest-dom';
 describe('App test', () => {
   it('should render app', () => {
     render(<App />);
-    expect(screen.getByText(/learn react/i)).toBeInTheDocument();
+    expect(screen.getByText(/Todo list/i)).toBeInTheDocument();
   });
 });
