--- src/TodoItem.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { ITask } from './Interfaces';
-
-interface TaskProps {
-  task: ITask;
-}
-
-const Todoitem = ({ task }: TaskProps): JSX.Element => {
-  return <div key={task.id}>{task.name}</div>;
-};
-
-export default Todoitem;
